fix display string of unsigned bitfields > 32 bits (typo)
svn path=/trunk/KDE/kdesdk/okteta/; revision=1284519

+++ kasten/controllers/view/structures/datatypes/primitive/bitfield/unsignedbitfielddatainformation.cpp
@@ -29,7 +29,7 @@ QString UnsignedBitfieldDataInformation::valueString() const
 {
     if (!mWasAbleToRead)
         return i18nc("invalid value (out of range)", "<invalid>");
-    return UIntDataInformation<quint32>::valueString(mValue.ulongValue);
+    return UIntDataInformation<quint64>::valueString(mValue.ulongValue);
 }
 
 QWidget* UnsignedBitfieldDataInformation::createEditWidget(QWidget* parent) const
